requirements.txt
Django
pytz
djangorestframework
pygments
httpie